+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.6.2
 
 * clear session buffer before it's full or exceeds the buffer size
+* fixed concurrent access to redis session cache
 
 ## 3.6.1